Rework Auto stuff using interface and abstract class
