+++ src/lesson1/Electric.java
@@ -1,9 +1,6 @@
 package lesson1;
 
-import lesson1.Auto;
-
-
-public class Electric extends Auto {
+public class Electric extends AbstractIAuto {
     public Electric(String make, String model, int year, String engine){
         super(make, model, year, engine);
     }
