+++ src/lesson1/IAuto.java
@@ -1,0 +1,7 @@
+package lesson1;
+
+interface IAuto {
+    void drive();
+
+    void displayInfo();
+}
